In neighborhoods.h, add comments concerning enum neighborhood_type_nums. Keep the numbers in the range [0, 255], at least for now. The numbers must be non-negative since they are used as indices into the neighborhood_type_names array. Neighborhood type is treated as an unsigned 8-bit int by hodge-podge.c and is stored as an unsigned 16-bit int attribute in the output HDF5 file.
In hodge-podge.c, explicitly cast EXT_MOORE to uint8 for comparison with
uint8 neighborhood_type variable and for printing with %hhu format.

+++ src/hodge-podge.c
@@ -147,9 +147,9 @@ int main (int argc, char **argv)
       fprintf (stderr, "Failed to parse arg 4, neighborhood type (0, 1, or 2)\n");
       fatal++;
    }
-   if (EXT_MOORE != neighborhood_type) {
+   if ((uint8)EXT_MOORE != neighborhood_type) {
       fprintf (stderr, "Sorry, only neighborhood type %hhu (%s) is supported"
-         " right now\n", EXT_MOORE, neighborhood_type_names [EXT_MOORE]);
+         " right now\n", (uint8)EXT_MOORE, neighborhood_type_names [EXT_MOORE]);
       fatal++;
    }
    if ( 1 != sscanf (argv[5], "%hu", &V) ) {
